+++ src/gamspy/_symbols/equation.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import builtins
 import itertools
 import uuid
-import warnings
 from enum import Enum
 from typing import TYPE_CHECKING, Any
 
@@ -274,7 +273,7 @@ class Equation(gt.Equation, operable.Operable, Symbol):
                 name = validation.validate_name(name)
 
                 if is_miro_output:
-                    name = name.lower()
+                    name = name.lower()  # type: ignore
             else:
                 name = "e" + str(uuid.uuid4()).replace("-", "_")
 
@@ -686,21 +685,6 @@ class Equation(gt.Equation, operable.Operable, Symbol):
             raise ValidationError("Equation is not defined!")
 
         return self._definition.getDeclaration()
-
-    def getStatement(self) -> str:
-        """
-        Statement of the Equation declaration
-
-        Returns
-        -------
-        str
-        """
-        warnings.warn(
-            "getStatement is going to be renamed in 0.12.5. Please use getDeclaration instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self.getDeclaration()
 
 
 def cast_type(type: str | EquationType) -> str: