+++ tests/test_gamspy.py
@@ -45,7 +45,7 @@ class GamspySuite(unittest.TestCase):
     def test_version(self):
         import gamspy
 
-        self.assertEqual(gamspy.__version__, "0.12.4")
+        self.assertEqual(gamspy.__version__, "0.12.5")
 
 
 class DocsSuite(unittest.TestCase):